# backend/src/models/task.py
from typing import Optional
from sqlmodel import Field, SQLModel
from datetime import datetime

class TaskBase(SQLModel):
    user_id: str = Field(index=True)
    title: str
    description: Optional[str] = None
    completed: bool = False

class Task(TaskBase, table=True):
    id: Optional[int] = Field(default=None, primary_key=True)
    created_at: datetime = Field(default_factory=datetime.utcnow)
    updated_at: datetime = Field(default_factory=datetime.utcnow)