# specs/002-full-stack-todo-app/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Todo Application API
  version: 1.0.0
  description: API for the Full-Stack Todo Application.
servers:
  - url: /api/{user_id} # Base URL, user_id will be appended by the client
paths:
  /tasks:
    get:
      summary: Get all tasks for the authenticated user.
      operationId: getAllTasks
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the authenticated user.
      responses:
        '200':
          description: A list of tasks.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized. Invalid or missing JWT.
    post:
      summary: Create a new task for the authenticated user.
      operationId: createTask
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the authenticated user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskInput'
      responses:
        '201':
          description: Task created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad Request. Invalid payload.
        '401':
          description: Unauthorized. Invalid or missing JWT.
  /tasks/{id}:
    get:
      summary: Get a specific task by ID for the authenticated user.
      operationId: getTaskById
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the authenticated user.
        - name: id
          in: path
          required: true
          schema:
            type: string # Assuming UUID or integer represented as string
          description: The ID of the task to retrieve.
      responses:
        '200':
          description: The requested task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized. Invalid or missing JWT.
        '404':
          description: Not Found. Task does not exist for this user.
    put:
      summary: Update an existing task for the authenticated user.
      operationId: updateTask
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the authenticated user.
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the task to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskInput' # Can be the same as create input or a different one for full update
      responses:
        '200':
          description: Task updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad Request. Invalid payload.
        '401':
          description: Unauthorized. Invalid or missing JWT.
        '404':
          description: Not Found. Task does not exist for this user.
    delete:
      summary: Delete a task for the authenticated user.
      operationId: deleteTask
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the authenticated user.
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the task to delete.
      responses:
        '204':
          description: Task deleted successfully. No content returned.
        '401':
          description: Unauthorized. Invalid or missing JWT.
        '404':
          description: Not Found. Task does not exist for this user.
  /tasks/{id}/complete:
    patch:
      summary: Toggle the completion status of a task for the authenticated user.
      operationId: toggleTaskCompletion
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the authenticated user.
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the task to toggle.
      responses:
        '200':
          description: Task completion status toggled successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized. Invalid or missing JWT.
        '404':
          description: Not Found. Task does not exist for this user.

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: string # Assuming UUID, string is safest
          description: Unique identifier for the task.
          example: "d290f1bb-2c8e-4146-838a-7c4e67f5c91f"
        user_id:
          type: string
          description: The ID of the user who owns this task.
          example: "user-123"
        title:
          type: string
          description: The title of the task.
          example: "Buy groceries"
        description:
          type: string
          description: A detailed description of the task.
          example: "Milk, Bread, Eggs, Cheese"
        completed:
          type: boolean
          description: Indicates if the task is completed.
          example: false
        created_at:
          type: string
          format: date-time
          description: Timestamp when the task was created.
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the task was last updated.
      required:
        - id
        - user_id
        - title
        - completed
        - created_at
        - updated_at
    TaskInput:
      type: object
      properties:
        title:
          type: string
          description: The title of the task.
          example: "Buy groceries"
        description:
          type: string
          description: A detailed description of the task.
          example: "Milk, Bread, Eggs, Cheese"
      required:
        - title